+++ adb-20190315/src/Models/CancelActiveOperationTasksResponseBody.php
@@ -1,0 +1,63 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Adb\V20190315\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class CancelActiveOperationTasksResponseBody extends Model
+{
+    /**
+     * @example am-xxxxxxxx
+     *
+     * @var string
+     */
+    public $ids;
+
+    /**
+     * @example 642F3512-C628-5D0C-8815-F6670CEA00D4
+     *
+     * @var string
+     */
+    public $requestId;
+    protected $_name = [
+        'ids'       => 'Ids',
+        'requestId' => 'RequestId',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->ids) {
+            $res['Ids'] = $this->ids;
+        }
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return CancelActiveOperationTasksResponseBody
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Ids'])) {
+            $model->ids = $map['Ids'];
+        }
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+
+        return $model;
+    }
+}