+++ adb-20190315/src/Models/DescribeLoghubDetailRequest.php
@@ -6,22 +6,16 @@ namespace AlibabaCloud\SDK\Adb\V20190315\Models;
 
 use AlibabaCloud\Tea\Model;
 
-class AttachUserENIRequest extends Model
+class DescribeLoghubDetailRequest extends Model
 {
     /**
-     * @var string
-     */
-    public $accessKeyIdCopy;
-
-    /**
-     * @description The ID of the AnalyticDB for MySQL Data Warehouse Edition (V3.0) cluster.
+     * @description This parameter is required.
      *
-     * This parameter is required.
-     * @example am-bp11q28kvl688****
+     * @example wap_log_full_to_adb
      *
      * @var string
      */
-    public $DBClusterId;
+    public $exportName;
 
     /**
      * @var string
@@ -34,6 +28,31 @@ class AttachUserENIRequest extends Model
     public $ownerId;
 
     /**
+     * @description This parameter is required.
+     *
+     * @example sls-hz-php-ad
+     *
+     * @var string
+     */
+    public $projectName;
+
+    /**
+     * @description This parameter is required.
+     *
+     * @example cn-hangzhou
+     *
+     * @var string
+     */
+    public $regionId;
+
+    /**
+     * @example rg-4690g37929****
+     *
+     * @var string
+     */
+    public $resourceGroupId;
+
+    /**
      * @var string
      */
     public $resourceOwnerAccount;
@@ -43,10 +62,12 @@ class AttachUserENIRequest extends Model
      */
     public $resourceOwnerId;
     protected $_name = [
-        'accessKeyIdCopy'      => 'AccessKeyId-copy',
-        'DBClusterId'          => 'DBClusterId',
+        'exportName'           => 'ExportName',
         'ownerAccount'         => 'OwnerAccount',
         'ownerId'              => 'OwnerId',
+        'projectName'          => 'ProjectName',
+        'regionId'             => 'RegionId',
+        'resourceGroupId'      => 'ResourceGroupId',
         'resourceOwnerAccount' => 'ResourceOwnerAccount',
         'resourceOwnerId'      => 'ResourceOwnerId',
     ];
@@ -58,17 +79,23 @@ class AttachUserENIRequest extends Model
     public function toMap()
     {
         $res = [];
-        if (null !== $this->accessKeyIdCopy) {
-            $res['AccessKeyId-copy'] = $this->accessKeyIdCopy;
-        }
-        if (null !== $this->DBClusterId) {
-            $res['DBClusterId'] = $this->DBClusterId;
+        if (null !== $this->exportName) {
+            $res['ExportName'] = $this->exportName;
         }
         if (null !== $this->ownerAccount) {
             $res['OwnerAccount'] = $this->ownerAccount;
         }
         if (null !== $this->ownerId) {
             $res['OwnerId'] = $this->ownerId;
+        }
+        if (null !== $this->projectName) {
+            $res['ProjectName'] = $this->projectName;
+        }
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+        if (null !== $this->resourceGroupId) {
+            $res['ResourceGroupId'] = $this->resourceGroupId;
         }
         if (null !== $this->resourceOwnerAccount) {
             $res['ResourceOwnerAccount'] = $this->resourceOwnerAccount;
@@ -83,22 +110,28 @@ class AttachUserENIRequest extends Model
     /**
      * @param array $map
      *
-     * @return AttachUserENIRequest
+     * @return DescribeLoghubDetailRequest
      */
     public static function fromMap($map = [])
     {
         $model = new self();
-        if (isset($map['AccessKeyId-copy'])) {
-            $model->accessKeyIdCopy = $map['AccessKeyId-copy'];
-        }
-        if (isset($map['DBClusterId'])) {
-            $model->DBClusterId = $map['DBClusterId'];
+        if (isset($map['ExportName'])) {
+            $model->exportName = $map['ExportName'];
         }
         if (isset($map['OwnerAccount'])) {
             $model->ownerAccount = $map['OwnerAccount'];
         }
         if (isset($map['OwnerId'])) {
             $model->ownerId = $map['OwnerId'];
+        }
+        if (isset($map['ProjectName'])) {
+            $model->projectName = $map['ProjectName'];
+        }
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+        if (isset($map['ResourceGroupId'])) {
+            $model->resourceGroupId = $map['ResourceGroupId'];
         }
         if (isset($map['ResourceOwnerAccount'])) {
             $model->resourceOwnerAccount = $map['ResourceOwnerAccount'];